Fix: JpaRepository타입 파라미터 변경 - Long -> Integer

+++ src/main/java/com/hrm/repository/RecruitmentPostRepository.java
@@ -9,7 +9,7 @@ import com.hrm.entity.RecruitmentPostEntity;
 import com.hrm.enums.PostStatus;
 
 @Repository
-public interface RecruitmentPostRepository extends JpaRepository<RecruitmentPostEntity, Long> {
+public interface RecruitmentPostRepository extends JpaRepository<RecruitmentPostEntity, Integer> {
 	// status 로 모든 채용공고를 찾는 메소드
 	List<RecruitmentPostEntity> findByStatus(PostStatus status);
 }
